# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для управления переговорками, расписаниями и бронями.",
        "title": "API бронирования переговорок",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/bookings/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает постраничный список всех броней. Только для роли admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Брони"
                ],
                "summary": "Список всех броней",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы (по умолчанию 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы (по умолчанию 20, максимум 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BookingsListResponse"
                        }
                    },
                    "400": {
                        "description": "invalid pagination parameters",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список будущих броней текущего пользователя. Только для роли user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Брони"
                ],
                "summary": "Мои брони",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BookingsResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookings/{bookingId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Отменяет бронь текущего пользователя. Операция идемпотентна. Только для роли user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Брони"
                ],
                "summary": "Отменить бронь",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID брони",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.BookingResponse"
                        }
                    },
                    "400": {
                        "description": "invalid UUID",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "cannot cancel another user's booking",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "booking not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новую переговорку. Только для роли admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Переговорки"
                ],
                "summary": "Создать переговорку",
                "parameters": [
                    {
                        "description": "Данные переговорки",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRoomResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/rooms/{roomId}/schedule/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт расписание доступности для переговорки. Расписание можно задать только один раз. Только для роли admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Создать расписание переговорки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID переговорки",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Расписание",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateScheduleResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "room not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "schedule already exists",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.Booking": {
            "type": "object",
            "properties": {
                "conferenceLink": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "slotId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "dto.BookingResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/dto.Booking"
                }
            }
        },
        "dto.BookingsListResponse": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Booking"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/dto.Pagination"
                }
            }
        },
        "dto.BookingsResponse": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Booking"
                    }
                }
            }
        },
...
        "dto.CreateRoomRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateRoomResponse": {
            "type": "object",
            "properties": {
                "room": {
                    "$ref": "#/definitions/dto.Room"
                }
            }
        },
        "dto.CreateScheduleRequest": {
            "type": "object",
            "required": [
                "endTime",
                "startTime"
            ],
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "dto.CreateScheduleResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "$ref": "#/definitions/dto.Schedule"
                }
            }
        },
...
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Room": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.Schedule": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "httpx.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/httpx.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT-токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}